% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_gg.R
\name{theme_campbead}
\alias{theme_campbead}
\title{Light colored campbead theme for ggplot}
\usage{
theme_campbead()
}
\description{
This function all sets a light-colored theme for ggplot
}
\examples{
theme_campbead()
}
\keyword{ggplot,}
\keyword{theme}
