% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_gg.R
\name{theme_campbead_dark}
\alias{theme_campbead_dark}
\title{Dark colored campbead theme for ggplot}
\usage{
theme_campbead_dark()
}
\description{
This function all sets a dark-colored theme for ggplot
}
\examples{
theme_campbead_dark()
}
\keyword{ggplot,}
\keyword{theme}
